Stop abusing EUMM too much
This code breaks on the EUMM version that'll be shipped with 5.14.0 due to a bug
in EUMM itself. However, if we stop doing crazy shit to prevent oldinterface.pod
to be installed, things will be fine. Also, at least recent versions of EUMM
won't install that extra pod file anyway. I haven't tested on older ones, but
I'm quite sure we don't care, as all that could go wrong is one extra file being
installed.